separated the spawning logic from the wave logic for expandability
