+++ .github/CODEOWNERS
@@ -26,3 +26,4 @@ SubTagSystem/ @domibron
 **/Util/Fun/ @domibron
 **/Util/RefGetters/ @domibron
 **/Util/RemoveObject/ @domibron
+Constants/ @domibron
