+++ Assets/Programming/Enemies/Spawn System/SpawnHandler.cs
@@ -1,0 +1,182 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+
+//by    _                 _ _                     
+//     | |               (_) |                    
+//   __| | ___  _ __ ___  _| |__  _ __ ___  _ __  
+//  / _` |/ _ \| '_ ` _ \| | '_ \| '__/ _ \| '_ \ 
+// | (_| | (_) | | | | | | | |_) | | | (_) | | | |
+//  \__,_|\___/|_| |_| |_|_|_.__/|_|  \___/|_| |_|
+// © 2025 Dominic McNeill dommcneill@outlook.com
+
+
+/// <summary>
+/// Base script to handle the spawning of enemies.
+/// </summary>
+public class SpawnHandler : MonoBehaviour
+{
+    /// <summary>
+    /// The scriptable object that stores the enemy prefabs and their corresponding numerical id.
+    /// </summary>
+    public EnemyDataSO enemyCoreData;
+
+    public List<GameObject> spawnPoints;
+
+
+    protected List<Transform> trackingEnemies = new List<Transform>();
+
+    protected int enemyCount = 0;
+
+
+    #region SpawnAI
+    /// <summary>
+    /// Spawn the AI with the given id, min radius to spawn and maxRadius.
+    /// </summary>
+    /// <param name="id">The id of the ai to look for in aiSpawnList.</param>
+    /// <param name="minRadius">How close the AI can spawn.</param>
+    /// <param name="maxRadius">How far the AI can spawn.</param>
+    /// <returns>True if spawning was successful.</returns>
+    protected bool SpawnAI(int id, List<GameObject> spawnLocationToUse, Vector3 playerPosition, float minRadius, float maxRadius)
+    {
+        // get the AI prefab from the id given.
+        GameObject aiPrefab = GetEnemyPrefabFromID(id);
+
+        if (aiPrefab == null)
+        {
+            Debug.LogWarning($"Prefab missing for ai id of {id}");
+            return false;
+        }
+
+        // get a random spawn location that is suitable.
+        Vector3? spawnLocation = GetRandomSpawnLocation(spawnLocationToUse, playerPosition, minRadius, maxRadius);
+
+        if (!spawnLocation.HasValue)
+        {
+            Debug.LogWarning($"Cannot find a suitable location to spawn the AI");
+            return false;
+        }
+
+        // spawn the AI and set state to alerted. This is not proc gen. this is wave gen.
+        GameObject ai = Instantiate(aiPrefab, spawnLocation.Value, Quaternion.identity, null);
+        ai.GetComponent<AIBase>().ChangeState(AIState.Alerted);
+        ai.GetComponent<AIBase>().onDeath += RemoveEnemy;
+
+        AddEnemy(ai.transform);
+
+        return true;
+    }
+    #endregion
+
+
+    #region GetRandomSpawnLocation
+    /// <summary>
+    /// Searches through the spawnLocations for points that are within min and max distances from the player and returns a random point.
+    /// </summary>
+    /// <param name="minRadius">How close the point can be to the player.</param>
+    /// <param name="maxRadius">How far the point can be from the player.</param>
+    /// <returns>A point as Vector3 if successful, null if not.</returns>
+    private Vector3? GetRandomSpawnLocation(List<GameObject> spawnLocationToUse, Vector3 playerPosition, float minRadius = 30f, float maxRadius = 100f)
+    {
+        List<GameObject> viableSpawnLocations = new List<GameObject>();
+
+        // we sore all possible spawn locations.
+        foreach (GameObject possibleSpawnLocation in spawnLocationToUse)
+        {
+            if (possibleSpawnLocation == null) continue;
+
+            if (Vector3.Distance(possibleSpawnLocation.transform.position, playerPosition) > minRadius &&
+                Vector3.Distance(possibleSpawnLocation.transform.position, playerPosition) < maxRadius)
+            {
+                viableSpawnLocations.Add(possibleSpawnLocation);
+            }
+            else
+            {
+                continue;
+            }
+        }
+
+        // if we have none, then we return null.
+        if (viableSpawnLocations.Count <= 0) return null;
+
+        // pick a random point from the collection.
+        return viableSpawnLocations[UnityEngine.Random.Range(0, viableSpawnLocations.Count)].transform.position;
+    }
+    #endregion
+
+
+    #region GetEnemyPrefabFromID
+    /// <summary>
+    /// Get the prefab from the id provided. Searches through aiSpawnList for that id and returns the prefab.
+    /// </summary>
+    /// <param name="id">The id of the AI in the aiSpawnList.</param>
+    /// <returns>Returns the prefab object if successful, null if there wasn't one.</returns>
+    private GameObject GetEnemyPrefabFromID(int id)
+    {
+        foreach (Enemy enemy in enemyCoreData.aISpawnList)
+        {
+            if (enemy.id == id) return enemy.prefab;
+        }
+
+        return null;
+
+    }
+    #endregion
+
+    #region RemoveEnemy
+    /// <summary>
+    /// Removes the enemy that died.
+    /// </summary>
+    /// <param name="entityTransform">The transform of the enemy that died, may be null.</param>
+    private void RemoveEnemy(Transform entityTransform)
+    {
+        enemyCount--;
+        trackingEnemies.Remove(entityTransform);
+    }
+    #endregion
+
+    #region AddEnemy
+    private void AddEnemy(Transform enemyTransform)
+    {
+        enemyCount++;
+        trackingEnemies.Add(enemyTransform);
+    }
+    #endregion
+
+    #region RandomiseList
+    /// <summary>
+    /// Util function to randomise a given list.
+    /// </summary>
+    /// <param name="list">The list to randomise.</param>
+    protected void RandomiseList(ref List<int> list)
+    {
+
+        for (int i = 0; i < list.Count; i++)
+        {
+            int itemA = list[i];
+            int randomLocation = UnityEngine.Random.Range(0, list.Count);
+
+            int itemB = list[randomLocation];
+            list[randomLocation] = itemA;
+
+            list[i] = itemB;
+        }
+
+    }
+    #endregion
+
+    protected List<int> ConvertEnemiesInWaveToIDList(EnemiesInWaveData[] enemyInWaveData)
+    {
+        // I swear there is a func or something built in to create a collection of the same value.
+        List<int> listToReturn = new List<int>();
+        foreach (EnemiesInWaveData enemy in enemyInWaveData)
+        {
+            for (int i = 0; i < enemy.amount; i++)
+            {
+                listToReturn.Add(enemy.id);
+            }
+        }
+        return listToReturn;
+    }
+}
